services/lua-validator/pyproject.toml
[project]
name = "lua-validator"
version = "0.1.0"
description = "gRPC service that validates Lua code by actually running it"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
]
